ShoppingList

+++ app/src/main/java/com/dzenis_ska/shopinglist/db/MainViewModel.kt
@@ -2,6 +2,7 @@ package com.dzenis_ska.shopinglist.db
 
 import android.util.Log
 import androidx.lifecycle.*
+import com.dzenis_ska.shopinglist.entities.LibraryItem
 import com.dzenis_ska.shopinglist.entities.NoteItem
 import com.dzenis_ska.shopinglist.entities.ShopListItem
 import com.dzenis_ska.shopinglist.entities.ShopListNameItem
@@ -11,6 +12,8 @@ import java.lang.IllegalArgumentException
 class MainViewModel(dataBase: MainDataBase) : ViewModel() {
     val dao = dataBase.getDao()
 
+    val libraryItems = MutableLiveData<List<LibraryItem>>()
+
     val allNotes : LiveData<List<NoteItem>> = dao.getAllNotes().asLiveData()
 
     val allShopListNamesItem : LiveData<List<ShopListNameItem>> = dao.getAllShopListNames().asLiveData()
@@ -18,6 +21,10 @@ class MainViewModel(dataBase: MainDataBase) : ViewModel() {
     fun getAllItemsFromList(listId: Int): LiveData<List<ShopListItem>>{
         return dao.getAllShopListItems(listId).asLiveData()
     }
+
+    fun getAllLibraryItems(name: String) = viewModelScope.launch {
+        libraryItems.value = dao.getAllLibraryItems(name)
+        }
 
     fun insertNote(note: NoteItem) {
         viewModelScope.launch {
@@ -30,10 +37,15 @@ class MainViewModel(dataBase: MainDataBase) : ViewModel() {
     fun insertShopItem(shopListItem: ShopListItem) = viewModelScope.launch {
 //        Log.d("!!!", "${shopListItem}")
         dao.insertItem(shopListItem)
+        if(!isLibraryItemExists(shopListItem.name)) dao.insertLibraryItem(LibraryItem(null, shopListItem.name))
     }
 
     fun updateNote(note: NoteItem) = viewModelScope.launch {
         dao.updateNote(note)
+    }
+
+    fun updateLibraryItem(item: LibraryItem) = viewModelScope.launch {
+        dao.updateLibraryItem(item)
     }
 
     fun updateListItem(note: ShopListItem) = viewModelScope.launch {
@@ -48,13 +60,22 @@ class MainViewModel(dataBase: MainDataBase) : ViewModel() {
             dao.deleteNote(id)
         }
 
-    fun deleteShopListName(id: Int) = viewModelScope.launch {
-        dao.deleteShopListName(id)
+    fun deleteLibraryItem(id: Int) = viewModelScope.launch {
+        dao.deleteLibraryItem(id)
+    }
+
+    fun deleteShopList(id: Int, deleteList: Boolean) = viewModelScope.launch {
+        if(deleteList) dao.deleteShopListName(id)
+        dao.deleteShopItemsById(id)
+    }
+
+    private suspend fun isLibraryItemExists(name: String): Boolean {
+        return dao.getAllLibraryItems(name).isNotEmpty()
     }
 
 
     class MainViewModelFactory(val dataBase: MainDataBase) : ViewModelProvider.Factory{
-        override fun <T : ViewModel?> create(modelClass: Class<T>): T {
+        override fun <T : ViewModel> create(modelClass: Class<T>): T {
                 if(modelClass.isAssignableFrom(MainViewModel::class.java)){
                     @Suppress("UNCHECKED_CAST")
                     return MainViewModel(dataBase) as T
